move model type
