+++ web/src/components/molecules/Content/List/index.tsx
@@ -9,8 +9,8 @@ import { ProColumns } from "@reearth-cms/components/atoms/ProTable";
 import Sidebar from "@reearth-cms/components/molecules/Common/Sidebar";
 import ContentTable from "@reearth-cms/components/molecules/Content/Table";
 import { ContentTableField, Item } from "@reearth-cms/components/molecules/Content/types";
+import { Model } from "@reearth-cms/components/molecules/Model/types";
 import { Request } from "@reearth-cms/components/molecules/Request/types";
-import { Model } from "@reearth-cms/components/molecules/Schema/types";
 import { ItemSort, AndConditionInput } from "@reearth-cms/components/molecules/View/types";
 import { CurrentViewType } from "@reearth-cms/components/organisms/Project/Content/ContentList/hooks";
 import { useT } from "@reearth-cms/i18n";
