+++ web/src/components/molecules/Schema/types.ts
@@ -1,15 +1,3 @@
-export type Model = {
-  id: string;
-  name: string;
-  description?: string;
-  key: string;
-  schemaId: string;
-  schema: Schema;
-  metadataSchema?: MetaDataSchema;
-  public: boolean;
-  order?: number;
-};
-
 export type MetaDataSchema = {
   id?: string;
   fields?: Field[];
