+++ web/src/components/organisms/Project/ModelsMenu/hooks.ts
@@ -2,7 +2,9 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 import Notification from "@reearth-cms/components/atoms/Notification";
-import { Model, Group, ModelFormValues } from "@reearth-cms/components/molecules/Schema/types";
+import { Model } from "@reearth-cms/components/molecules/Model/types";
+import { Group, ModelFormValues } from "@reearth-cms/components/molecules/Schema/types";
+import { fromGraphQLModel } from "@reearth-cms/components/organisms/DataConverters/model";
 import {
   useGetModelsQuery,
   useGetGroupsQuery,
@@ -16,7 +18,7 @@ import {
 } from "@reearth-cms/gql/graphql-client-api";
 import { useT } from "@reearth-cms/i18n";
 import { useModel, useWorkspace, useProject } from "@reearth-cms/state";
-import { fromGraphQLModel, fromGraphQLGroup } from "@reearth-cms/utils/values";
+import { fromGraphQLGroup } from "@reearth-cms/utils/values";
 
 type Params = {
   modelId?: string;
