+++ web/src/components/organisms/Project/Schema/hooks.ts
@@ -3,9 +3,11 @@ import { useCallback, useEffect, useMemo, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 
 import Notification from "@reearth-cms/components/atoms/Notification";
+import { Model } from "@reearth-cms/components/molecules/Model/types";
 import { SelectedSchemaType } from "@reearth-cms/components/molecules/Schema";
-import { Field, FieldType, Model, Group } from "@reearth-cms/components/molecules/Schema/types";
+import { Field, FieldType, Group } from "@reearth-cms/components/molecules/Schema/types";
 import type { FormValues, ModelFormValues } from "@reearth-cms/components/molecules/Schema/types";
+import { fromGraphQLModel } from "@reearth-cms/components/organisms/DataConverters/model";
 import {
   useCreateFieldMutation,
   SchemaFieldType,
@@ -28,7 +30,7 @@ import {
 } from "@reearth-cms/gql/graphql-client-api";
 import { useT } from "@reearth-cms/i18n";
 import { useModel } from "@reearth-cms/state";
-import { fromGraphQLModel, fromGraphQLGroup } from "@reearth-cms/utils/values";
+import { fromGraphQLGroup } from "@reearth-cms/utils/values";
 
 export default () => {
   const t = useT();
