+++ web/src/state/index.ts
@@ -2,7 +2,7 @@ import { atom, useAtom } from "jotai";
 import { atomWithStorage } from "jotai/utils";
 
 import { PublicScope } from "@reearth-cms/components/molecules/Accessibility/types";
-import { Model } from "@reearth-cms/components/molecules/Schema/types";
+import { Model } from "@reearth-cms/components/molecules/Model/types";
 import { Role } from "@reearth-cms/components/molecules/Workspace/types";
 
 export type Workspace = {
