+++ web/src/utils/values.ts
@@ -1,4 +1,4 @@
-import { Model, Field, Group } from "@reearth-cms/components/molecules/Schema/types";
+import { Field, Group } from "@reearth-cms/components/molecules/Schema/types";
 import {
   View,
   AndConditionInput,
@@ -8,7 +8,6 @@ import {
 } from "@reearth-cms/components/molecules/View/types";
 import {
   Maybe,
-  Model as GQLModel,
   Group as GQLGroup,
   View as GQLView,
   FieldType as GQLFieldType,
@@ -17,56 +16,6 @@ import {
   ItemSortInput as GQLItemSortInput,
   ColumnSelectionInput as GQLColumnSelectionInput,
 } from "@reearth-cms/gql/graphql-client-api";
-
-export const fromGraphQLModel = (model: Maybe<GQLModel>): Model | undefined => {
-  if (!model) return;
-
-  return {
-    id: model.id,
-    description: model.description,
-    name: model.name,
-    key: model.key,
-    public: model.public,
-    order: model.order ?? undefined,
-    schema: {
-      id: model.schema?.id,
-      fields: model.schema?.fields.map(
-        field =>
-          ({
-            id: field.id,
-            description: field.description,
-            title: field.title,
-            type: field.type,
-            key: field.key,
-            unique: field.unique,
-            isTitle: field.isTitle,
-            multiple: field.multiple,
-            required: field.required,
-            typeProperty: field.typeProperty,
-          }) as Field,
-      ),
-    },
-    schemaId: model.schemaId,
-    metadataSchema: {
-      id: model.metadataSchema?.id,
-      fields: model.metadataSchema?.fields.map(
-        field =>
-          ({
-            id: field.id,
-            description: field.description,
-            title: field.title,
-            type: field.type,
-            key: field.key,
-            unique: field.unique,
-            isTitle: field.isTitle,
-            multiple: field.multiple,
-            required: field.required,
-            typeProperty: field.typeProperty,
-          }) as Field,
-      ),
-    },
-  };
-};
 
 export const fromGraphQLGroup = (group: Maybe<GQLGroup>): Group | undefined => {
   if (!group) return;
